not dev

+++ ci/deploy_conda.Jenkinsfile
@@ -135,9 +135,13 @@ pipeline {
             sh '''#!/bin/bash
               ${QLOGIN_ACTIVATE} &&
               echo ${JOBMON_VERSION}
-              rm -rf /ihme/centralcomp/docs/jobmon/zzzz_test/${JOBMON_VERSION}
-              cp -r /mnt/team/scicomp/pub/docs_temp /ihme/centralcomp/docs/jobmon/zzzz_test/${JOBMON_VERSION}
-              python $WORKSPACE/ci/publish.py '/ihme/centralcomp/docs/jobmon/zzzz_test'
+              if [[ "$JOBMON_VERSION" =~ "dev" ]]
+              then
+              else
+                rm -rf /ihme/centralcomp/docs/jobmon/zzzz_test/${JOBMON_VERSION}
+                cp -r /mnt/team/scicomp/pub/docs_temp /ihme/centralcomp/docs/jobmon/zzzz_test/${JOBMON_VERSION}
+                python $WORKSPACE/ci/publish.py '/ihme/centralcomp/docs/jobmon/zzzz_test'
+              fi
             '''
           } // end script
         } // end qlogin
